requirements.txt
flask
numpy
pillow
face_recognition
opencv-python-headless
mysql-connector-python
